# [DCJ-647] Optimize dataset_relationship retrieval