@okotsopoulos 2024-09-05
Closing the loop now that this has been released to prod, and the `dataset_table` index added.

New query with index **(cost = 864.67)**:

```
Nested Loop  (cost=0.29..864.67 rows=2 width=131) (actual time=0.150..3.021 rows=4 loops=1)
  ->  Index Scan using dataset_table_pkey on dataset_table dt  (cost=0.29..8.31 rows=1 width=16) (actual time=0.030..0.033 rows=1 loops=1)
        Index Cond: (id = 'a50a0be5-085b-4181-8df2-804c2298cb62'::uuid)
        Filter: (dataset_id = '9c259486-2ee8-4e1e-8d87-1741b1432194'::uuid)
  ->  Seq Scan on dataset_relationship r  (cost=0.00..856.34 rows=2 width=131) (actual time=0.118..2.985 rows=4 loops=1)
        Filter: (from_table = 'a50a0be5-085b-4181-8df2-804c2298cb62'::uuid)
        Rows Removed by Filter: 23512
Planning Time: 0.662 ms
Execution Time: 3.096 ms
```

So, a roughly **93% reduction in query cost** :tada: